Move block_state_external_stubbed into utils submodule, move unit test to plt_block_state_stub
