+++ plt/plt-scheduler/tests/plt_token_admin_role.rs
@@ -1,8 +1,5 @@
 //! Tests for token RBAC admin role operations via the scheduler.
 
-use crate::block_state_external_stubbed::{
-    BlockStateWithExternalStateStubbed, TokenInitTestParams,
-};
 use assert_matches::assert_matches;
 use concordium_base::base::{Energy, ProtocolVersion};
 use concordium_base::common::cbor;
@@ -13,8 +10,10 @@ use concordium_base::protocol_level_tokens::{
 use concordium_base::transactions::Payload;
 use plt_scheduler::{queries, scheduler};
 use plt_scheduler_types::types::execution::TransactionOutcome;
+use utils::block_state_external_stubbed::{
+    BlockStateWithExternalStateStubbed, TokenInitTestParams,
+};
 
-mod block_state_external_stubbed;
 mod utils;
 
 /// The governance account receives every role except for disabled features.
