+++ plt/plt-scheduler/tests/plt_token_list_updates.rs
@@ -1,8 +1,5 @@
 //! Tests for token allow/deny list update operations via the scheduler.
 
-use crate::block_state_external_stubbed::{
-    BlockStateWithExternalStateStubbed, TokenInitTestParams,
-};
 use assert_matches::assert_matches;
 use concordium_base::base::{AccountIndex, Energy};
 use concordium_base::common::cbor;
@@ -16,8 +13,10 @@ use concordium_base::transactions::Payload;
 use plt_scheduler::{queries, scheduler};
 use plt_scheduler_types::types::events::BlockItemEvent;
 use plt_scheduler_types::types::execution::TransactionOutcome;
+use utils::block_state_external_stubbed::{
+    BlockStateWithExternalStateStubbed, TokenInitTestParams,
+};
 
-mod block_state_external_stubbed;
 mod utils;
 
 /// Helper to execute a list operation as governance and return events.
