+++ plt/plt-scheduler/tests/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod block_state_external_stubbed;
+
 use assert_matches::assert_matches;
 use concordium_base::base::ProtocolVersion;
 use concordium_base::protocol_level_tokens::{
